+++ personnal/src/main/java/cgy/controller/DepartmentController.java
@@ -1,6 +1,7 @@
 package cgy.controller;
 
 import cgy.model.Department;
+import cgy.model.Employee;
 import cgy.model.Page;
 import cgy.model.Position;
 import cgy.service.DepartmentService;
@@ -28,6 +29,11 @@ public class DepartmentController {
     //管理员添加部门信息
     @RequestMapping("insertDepart")
     public String insertDepart(Department department, HttpServletRequest request) {
+        Employee employee = (Employee) request.getSession().getAttribute("employeeNow");
+        if (employee == null){
+            request.setAttribute("message","请先登录");
+            return "../index";
+        }
         department.setDep_date(new Date());
         boolean insertDepart = departmentService.insertDepart(department);
         //给出是否添加成功的判别信息
@@ -38,6 +44,11 @@ public class DepartmentController {
     //查看部门信息
     @RequestMapping("getDepart")
     public String getDepart(HttpServletRequest request) {
+        Employee employee = (Employee) request.getSession().getAttribute("employeeNow");
+        if (employee == null){
+            request.setAttribute("message","请先登录");
+            return "../index";
+        }
         String pageNoStr = request.getParameter("pageNo");
         int pageNo = 1;
         if (pageNoStr != null && !"".equals(pageNoStr)) {
@@ -54,6 +65,11 @@ public class DepartmentController {
     //去更新界面
     @RequestMapping("toupdateDepart")
     public String toupdateDepart(Department department, HttpServletRequest request) {
+        Employee employee = (Employee) request.getSession().getAttribute("employeeNow");
+        if (employee == null){
+            request.setAttribute("message","请先登录");
+            return "../index";
+        }
         //给出是否添加成功的判别信息
         request.setAttribute("department", department);
         return "admin/departmentupdate";
@@ -62,6 +78,11 @@ public class DepartmentController {
     //管理员更新部门信息
     @RequestMapping("updateDepart")
     public String updateDepart(Department department, HttpServletRequest request) {
+        Employee employee = (Employee) request.getSession().getAttribute("employeeNow");
+        if (employee == null){
+            request.setAttribute("message","请先登录");
+            return "../index";
+        }
         boolean updateDepart = departmentService.updateDepart(department);
         //给出是否添加成功的判别信息
         request.setAttribute("updateDepart", updateDepart);
@@ -71,6 +92,11 @@ public class DepartmentController {
     //管理员删除部门信息
     @RequestMapping("deleteDepart")
     public String deleteDepart(int dep_id, HttpServletRequest request) {
+        Employee employee = (Employee) request.getSession().getAttribute("employeeNow");
+        if (employee == null){
+            request.setAttribute("message","请先登录");
+            return "../index";
+        }
         boolean deleteDepart = departmentService.deleteDepart(dep_id);
         request.setAttribute("deleteDepart", deleteDepart);
         return "forward:getDepart";
@@ -96,6 +122,11 @@ public class DepartmentController {
 
     @RequestMapping("toaddrecruit")
     public String toaddrecruit(HttpServletRequest request) {
+        Employee employee = (Employee) request.getSession().getAttribute("employeeNow");
+        if (employee == null){
+            request.setAttribute("message","请先登录");
+            return "../index";
+        }
         List<Department> departs = departmentService.getDeparts();
         request.setAttribute("departs", departs);
         List<Position> positions = positionService.getPositionByDep_id(null);
